mcp-servers/email-server/requirements.txt
mcp>=1.0.0
google-api-python-client==2.186.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3

